db: remove #[allow(dead_code)] and fix warnings
